Updating search index removal to fix broken search server.

+++ apps/search/models.py
@@ -140,14 +140,15 @@ class MUserSearch(mongo.Document):
             feed.index_stories_for_search()
         
     @classmethod
-    def remove_all(cls):
+    def remove_all(cls, drop_index=False):
         user_searches = cls.objects.all()
         logging.info(" ---> ~SN~FRRemoving ~SB%s~SN user searches..." % user_searches.count())
         for user_search in user_searches:
             user_search.remove()
-            
-        logging.info(" ---> ~FRRemoving stories search index...")
-        SearchStory.drop()
+        
+        if drop_index:
+            logging.info(" ---> ~FRRemoving stories search index...")
+            SearchStory.drop()
         
     def remove(self):
         from apps.rss_feeds.models import Feed
@@ -388,6 +389,7 @@ class SearchFeed:
         
     @classmethod
     def query(cls, text):
+        cls.create_elasticsearch_mapping()
         try:
             cls.ES.default_indices = cls.index_name()
             cls.ES.indices.refresh()
